@@ -6,7 +6,7 @@ publish_to: 'none'
 version: 1.0.0+1
 
 environment:
-  sdk: '>=2.18.4 <3.0.0'
+  sdk: '>=3.0.0 <=3.13.3'
 
 dependencies:
   flutter:
@@ -18,6 +18,7 @@ dependencies:
   json_serializable: ^6.5.4
   json_annotation: ^4.7.0
   dio: ^4.0.6
+  provider: ^6.0.5
 
 dev_dependencies:
   flutter_test:
@@ -29,4 +30,3 @@ flutter:
   uses-material-design: true
   assets:
     - assets/
-    - assets/items_images/
